% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_allele_freq.R
\name{calculateAlleleFreq}
\alias{calculateAlleleFreq}
\title{XIBD Calculate Population Allele Frequencies}
\usage{
calculateAlleleFreq(ped.genotypes)
}
\arguments{
\item{ped.genotypes}{a named list containing \code{pedigree}, \code{genotypes} and \code{model}.
See \code{Value} description in \code{\link{getGenotypes}} for more details.}
}
\value{
a data frame with information:
\enumerate{
\item Chromosome (type \code{"character"}, \code{"numeric"} or \code{"integer"})
\item SNP identifiers (type \code{"character"})
\item Genetic map distance (Morgans, M) (type \code{"numeric"})
\item Base-pair position (type \code{"numeric"} or \code{"integer"})
\item Population allele frequency (type \code{"numeric"})
}
where each row defines a unique SNP.
Allele frequencies of -1 indicate missing data for all samples for that SNP.
}
\description{
Calculate population allele frequencies from genotypes.
}
