% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_pedmap.R
\docType{data}
\name{example_pedmap}
\alias{example_pedmap}
\title{Simulated IBD Data for Example}
\format{List with two objects
\describe{
\item{PED}{PLINK PED data frame. See \code{\link{getGenotypes}} for details.}
\item{MAP}{PLINK MAP data frame. See \code{\link{getGenotypes}} for details.}
}}
\usage{
example_pedmap
}
\description{
Simulated haplotype data in PLINK PED/MAP format for 28,808 SNPs
from chromosomes 1-23.
The data is for 10 samples from a 5 generation family with the most
distant pair being third cousins and the shortest distant pair being
parent-offspring.
The data was simulated for SNPs from an Illumina SNPchip under Illuminas
TOPBOT allele naming convention (see \code{switchBOTgenotypes}).
}
\keyword{datasets}
