% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_palette.R
\name{getColourPaletteMajor}
\alias{getColourPaletteMajor}
\title{Internal Function}
\usage{
getColourPaletteMajor(number.colours)
}
\arguments{
\item{number.colours}{numeric. The number of colours to return.}
}
\value{
A character vector of length=\code{number.colours} containing a colour specturm.
}
\description{
XIBD Colour Palette
}
\details{
Generate a spectrum colour palette with a specified number of colours.
}
