% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_palette.R
\name{getColourPaletteMinor}
\alias{getColourPaletteMinor}
\title{Generate a specified number of shades of a given colour.}
\usage{
getColourPaletteMinor(major.colour, number.colours)
}
\arguments{
\item{major.colour}{character. The colour name or code.}

\item{number.colours}{numeric. The number of colours to return.}
}
\value{
A character vector of length=\code{number.colours} containing colour shades, excluding white.
}
\description{
Generate a specified number of shades of a given colour.
}
