% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition_snps.R
\name{getConditionSNP}
\alias{getConditionSNP}
\title{Internal Function}
\usage{
getConditionSNP(genotypes, snp.ld, maximum.ld.r2)
}
\arguments{
\item{genotypes}{A data frame containing genotype information for a set of SNPs.
See \code{\link{getGenotypes}} for more details.}

\item{snp.ld}{A data frame generated from PLINK containing information on LD between pairs
of SNPs (A and B). This data frame contains exactly 7 columns of informations:
\enumerate{
\item Chromosome of SNP A (type \code{"numeric"} or \code{"integer"})
\item Base-pair position of SNP A (type \code{"numeric"} or \code{"integer"})
\item SNP A identifier (type \code{"character"})
\item Chromosome of SNP B (type \code{"numeric"} or \code{"integer"})
\item Base-pair position of SNP B (type \code{"numeric"} or \code{"integer"})
\item SNP B identifier (type \code{"character"})
\item R-sequared LD statistic between SNP A and SNP B (type \code{"numeric"})
}
where each row contains the LD information for a single pair of SNPs. The data frame should contain the header
\code{colnames(snp.ld)=c(CHR_A, BP_A, SNP_A, CHR_B, BP_B, SNP_B, R2)}}

\item{maximum.ld.r2}{A numeric value denoting the maximum linkage disequilibrium R2 value allowed
between pairs of SNPs.}
}
\value{
A data frame containing columns
\enumerate{
\item Chromosome (type \code{numeric} or \code{integer})
\item SNP identifier (type \code{character})
\item Genetic map position (type \code{numeric} or \code{integer})
\item Base-pair position (type \code{numeric} or \code{integer})
\item Numeric SNP identifier (type \code{numeric} or \code{integer})
\item Numeric condition SNP identifier (type \code{numeric} or \code{integer})
}
with columns headed \code{chr, snp_id, pos_M, pos_bp, marker.id} and \code{cond.snp.id}.
}
\description{
IBD Summary Table
}
\details{
Extracts the SNP with the largest LD with the current SNP
and returns this as the condition SNP
}
