% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{getGenotypes}
\alias{getGenotypes}
\title{XIBD Pre-Analysis Data Processing}
\usage{
getGenotypes(ped.map, reference.ped.map = NULL, snp.ld = NULL, model = 1,
  maf = 0.01, sample.max.missing = 0.1, snp.max.missing = 0.1,
  maximum.ld.r2 = 0.99, chromosomes = NULL, input.map.distance = "M",
  reference.map.distance = "M")
}
\arguments{
\item{ped.map}{a list with 2 objects:
\enumerate{
\item a data frame which contains the PLINK PED information.
The first six columns of this data frame are:
\enumerate{
\item Family ID (type \code{"character"}, \code{"numeric"} or \code{"integer"})
\item Individual ID (type \code{"character"}, \code{"numeric"} or \code{"integer"})
\item Paternal ID (type \code{"character"}, \code{"numeric"} or \code{"integer"})
\item Maternal ID (type \code{"character"}, \code{"numeric"} or \code{"integer"})
\item Gender (1 = male, 2 = female)
\item Phenotype (1 = unaffected, 2 = affected, 0 = unknown)
}
where each row describes a single sample The IDs are alphanumeric: the combination of family and individual ID should uniquely identify a sample. The phenotype column is not used
in XIBD analyses however it is required for completeness of a standard pedigree.
Columns 7 onwards are the sample haplotypes where the A and B alleles are coded as 1 and 2 respectively and missing data is coded as 0.
All SNPs (whether haploid or not) must have two alleles specified and each allele should be in a separate column. For example,
the alleles in columns 7 and 8 correspond to the unphased haplotypes of SNP 1 in the map file.
For haploid chromosomes, haplotypes should be specified as homozygous. Either both alleles should be missing (i.e. 0)
or neither. No header row should be given.
\item a data frame which contains the PLINK MAP information. This data frame contains exactly four columns of information:
\enumerate{
\item Chromosome (\code{"numeric"} or \code{"integer"})
\item SNP identifier (type \code{"character"})
\item Genetic map distance (centi morgans cM, or morgans M - default) (type \code{"numeric"})
\item Base-pair position (type \code{"numeric"} or \code{"integer"})
}
where each row describes a single marker. Genetic map distance and base-pair positions are expected to be positive values. The MAP file must
be ordered by increasing chromosomes and positions. SNP identifiers can contain any characters expect spaces or tabs; also you should avoid
* symbols in the names. The MAP file must contain as many markers as are in the PED file. No header row should be given.
}}

\item{reference.ped.map}{a list containing reference data used to calculate population allele
frequencies and haplotype frequencies, in the same format as \code{ped.map}.
The default value is \code{reference.ped.map=NULL} and XIBD will calculate the population allele
frequencies and haplotype frequencies from \code{ped.map}. This is not recommended for small datasets or
datasets of mixed populations. Genetic map positions and base-pair positions in this dataset are used
in the analysis. HapMap phase 2 and 3 PED and MAP data (hg19/build 37) for the 11 HapMap populations
can be downloaded from \url{http://bioinf.wehi.edu.au/software/XIBD}.}

\item{snp.ld}{optional for \code{model=1}; compulsory for \code{model=2}.
A data frame generated from PLINK containing information on LD between pairs of SNPs (A and B). This
data frame contains exactly 7 columns of information:
\enumerate{
\item Chromosome of SNP A (type \code{"numeric"} or \code{"integer"})
\item Base-pair position of SNP A (type \code{"numeric"} or \code{"integer"})
\item SNP A identifier (type \code{"character"})
\item Chromosome of SNP B (type \code{"numeric"} or \code{"integer"})
\item Base-pair position of SNP B (type \code{"numeric"} or \code{"integer"})
\item SNP B identifier (type \code{"character"})
\item R-squared LD statistic between SNP A and SNP B (type \code{"numeric"})
}
where each row contains the LD information for a single pair of SNPs. The data frame should contain the header
\code{CHR_A, BP_A, SNP_A, CHR_B, BP_B, SNP_B} and \code{R2}.
HapMap phase 2 and 3 PED and MAP data (hg19/build 37) for the 11 HapMap populations
can be downloaded from \url{http://bioinf.wehi.edu.au/software/XIBD}.
Alternatively, an LD file can be created using using PLINK (\url{http://www.cog-genomics.org/plink2}).}

\item{model}{an integer of either 1 or 2 denoting which of the two models should be run.
\enumerate{
\item \code{model=1} is based on the HMM implemented in PLINK (Purcell et al., 2007) which assumes the SNPs
are in linkage equilibrium (LE). This often requires thinning of markers prior to use.
\item \code{model=2} is based on the HMM implemented in RELATE (Albrechtsen et al., 2009) which allows SNPs
to be in LD and implicitly accounts for the LD through conditional emission probabilities where
the current genotype probability is conditioned on the genotype of a single previous SNP (haplotype frequencies).
}
\code{model=1} requires significantly less time to run than \code{model=2} due to the reduced number of SNPs
and simplified emission probabilities. It may be wise to use \code{model=1} when there are many SNPs and many
samples.}

\item{maf}{the smallest minor allele frequency allowed in the analysis. The
default value is 0.01.}

\item{sample.max.missing}{the maximum proportion of missing data allowed for
each sample. The default value is 0.1.}

\item{snp.max.missing}{the maximum proportion of missing data allowed for each
SNP. The default value is 0.1.}

\item{maximum.ld.r2}{the maximum linkage disequilibrium R2 value allowed
between pairs of SNPs.The default value is 0.99.}

\item{chromosomes}{a numeric vector containing a subset of chromosomes to perform genotype filtering on. The
default is \code{chromosomes=NULL} which will format genotypes for all chromosomes in \code{ped.map}.
Autosomes are represented by numbers 1-22 and the X chromosome is denoted 23.}

\item{input.map.distance}{either "M" or "cM" denoting whether the genetic map distances in
\code{ped.map} are in Morgans (M) or centi-Morgans (cM). The default is Morgans.}

\item{reference.map.distance}{either "M" or "cM" denoting whether the genetic map distances in
\code{reference.ped.map} are in Morgans (M) or centi-Morgans (cM). The default is Morgans. HapMap reference data is in Morgans.}
}
\value{
A named list of three objects:
\enumerate{
\item A pedigree containing the samples that remain after filtering. The pedigree is the first six columns
of the PED file and these columns are
headed \code{fid, iid, pid, mid, sex} and \code{aff}, respectively.
\item A data frame with the first five columns:
\enumerate{
\item Chromosome (type \code{"character"}, \code{"numeric"} or \code{"integer"})
\item SNP identifiers (type \code{"character"})
\item Genetic map distance (Morgans, M) (type \code{"numeric"})
\item Base-pair position (type \code{"numeric"} or \code{"integer"})
\item Population allele frequency (type \code{"numeric"})
}
where each row describes a single marker. These columns are headed \code{chr, snp_id, pos_M, pos_bp} and \code{freq} respectively.
If \code{model=2} then the following columns are also included:
\enumerate{
\item Numeric ID of condition SNP (type \code{"numeric"} or \code{"integer"})
\item Haplotype probability: pba (type \code{"numeric"})
\item Haplotype probability: pbA (type \code{"numeric"})
\item Haplotype probability: pBa (type \code{"numeric"})
\item Haplotype probability: pBA (type \code{"numeric"})
\item Population allele frequency on the condition SNP (type \code{"numeric"})
}
with the headers \code{condition_snp, pba, pbA, pBa, pBA} and \code{freq_condition_snp}.
The remaining columns contain the genotype data for each sample, where a single column corresponds to a single sample. These columns are
labeled with merged family IDs and individual IDs separated by a slash symbol (/).
\item The model selected.
}
The list is named \code{pedigree, genotypes} and \code{model} respectively.
}
\description{
Perform pre-analysis data processing of PLINK formatted unphased haplotype data,
including removal of SNPs and samples with high proportions of missing data, SNPs with low minor
allele frequencies and SNPs in high linkage disequilibrium (LD, based on R^2 if \code{model=2}).
Also, calculate population allele frequencies as well as haplotype frequencies between pairs of
SNPs (based on R^2 if \code{model=2}).
}
\examples{
# look at the simulated data
str(example_pedmap)

# format and filter the example data using model 2 and reference data
my_genotypes <- getGenotypes(ped.map = example_pedmap,
                             reference.ped.map = example_reference_pedmap,
                             snp.ld = example_reference_ld,
                             model = 2,
                             maf = 0.01,
                             sample.max.missing = 0.1,
                             snp.max.missing = 0.1,
                             maximum.ld.r2 = 0.99,
                             chromosomes = NULL,
                             input.map.distance = "M",
                             reference.map.distance = "M")
}
