% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition_snps.R
\name{getHaplotypeFreq}
\alias{getHaplotypeFreq}
\title{Calculates haplotype frequencies between two SNPs
using cpp scripts.}
\usage{
getHaplotypeFreq(genotypes, condition.snps, genders)
}
\arguments{
\item{genotypes}{A data frame containing genotype information for a set of SNPs.
See \code{\link{getGenotypes}} for more details.}

\item{condition.snps}{A dataframe containging information on SNPs to condition on.
See \code{\link{getConditionSNP}} for more details.}

\item{genders}{A numeric vector of the genders of the individuals in the genotypes
data frame. 1 = male, 2 = female.}
}
\value{
see return value from \code{\link{getGenotypes}} for more details.
}
\description{
Calculates haplotype frequencies between two SNPs
using cpp scripts.
}
