% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_estimates.R
\name{getIBDparameters}
\alias{getIBDparameters}
\title{XIBD Parameter Estimation}
\usage{
getIBDparameters(ped.genotypes, number.cores = 1)
}
\arguments{
\item{ped.genotypes}{a named list containing \code{pedigree}, \code{genotypes} and \code{model}.
See \code{Value} description in \code{\link{getGenotypes}} for more details.
The family IDs and individual IDs in \code{pedigree} must match the family IDs and individual IDs in the header of \code{genotypes}.}

\item{number.cores}{the number of cores used for parallel execution.}
}
\value{
A named list containing \code{autosome_parameters} and \code{X_chromosome_parameters}.
Both \code{autosome_parameters} and \code{X_chromosome_parameters} are data frames with columns:
\enumerate{
\item Family 1 ID (type \code{"character"})
\item Individual 1 ID (type \code{"character"})
\item Family 2 ID (type \code{"character"})
\item Individual 2 ID (type \code{"character"})
\item The number of meiosis (type \code{"numeric"} or \code{"integer"})
\item Probability of sharing 0 alleles IBD (type \code{"numeric"})
\item Probability of sharing 1 allele IBD (type \code{"numeric"})
\item Probability of sharing 2 alleles IBD (type \code{"numeric"})
}
headed \code{fid1, iid1, fid2, iid2, m, ibd0, ibd1} and \code{ibd2}, respectively.
Each row describes parameters for a unique pair of samples.
\code{autosome_parameters} will be \code{NULL} if chromosomes 1-22 are all excluded
from \code{ped.genotypes} and \code{X_chromosome_parameters} will be \code{NULL} if
chromosome 23 is excluded from \code{ped.genotypes}.
}
\description{
Estimate the number of meioses and the probabilities of sharing 0, 1 and 2 alleles IBD between pairs.
}
\examples{
# estimate parameters
my_parameters <- getIBDparameters(ped.genotypes = example_genotypes,
                                  number.cores = 1)

str(my_parameters)
}
