% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibd_detection.R
\name{getIBDsegments}
\alias{getIBDsegments}
\title{XIBDs IBD Segment Detection}
\usage{
getIBDsegments(ped.genotypes, parameters, model = NULL, chromosomes = NULL,
  number.cores = 1, minimum.snps = 20, minimum.length.bp = 50000,
  error = 0.001, posterior = FALSE)
}
\arguments{
\item{ped.genotypes}{a named list containing \code{pedigree}, \code{genotypes} and \code{model}.
See \code{Value} description in \code{\link{getGenotypes}} for more details.
Note the family IDs and individual IDs in \code{pedigree} must match the family IDs and individual IDs in the header of \code{genotypes}.}

\item{parameters}{a data frame containing meioses and IBD probability estimates for all pairwise combinations of samples.
See \code{Value} description in \code{\link{getIBDparameters}} for more details.}

\item{model}{an integer of either 1 or 2 denoting which of the two models should be run.
\enumerate{
\item \code{model=1} is based on the HMM implemented in PLINK (Purcell et al., 2007) which assumes the SNPs
are in linkage equilibrium (LE). This often requires thinning of markers prior to use.
\item \code{model=2} is based on the HMM implemented in RELATE (Albrechtsen et al., 2009) which allows SNPs
to be in LD and implicitly accounts for the LD through conditional emission probabilities where
the current genotype probability is conditioned on the genotype of a single previous SNP.
}
The default is \code{model=NULL} and the model listed in \code{ped.genotypes} is used. NOTE: \code{model=2} can
only be used if \code{model=2} is in \code{ped.genotypes}.}

\item{chromosomes}{a numeric vector containing a subset of chromosomes to perform IBD analysis on. The
default is \code{chromosomes=NULL} and IBD analysis will be performed on all chromosomes in \code{ped.genotypes}.}

\item{number.cores}{the number of cores used for parallel execution.}

\item{minimum.snps}{the minimum number of SNPs in an IBD segment for it to be reported. The default value is 20 SNPs.}

\item{minimum.length.bp}{the minimum length of a reported IBD segment. The default value is 50,000 bp.}

\item{error}{the genotyping error rate. The default value is 0.001.}

\item{posterior}{a logical value indicating whether posterior probabilities for each pairwise analysis should be returned.
The posterior probability is calculated for each SNP as posteriorPr(IBD=1)/2 + posteriorPr(IBD=2) using the
forward and backward variables (Rabiner, 1989).
A data frame containing probabilities for each SNP and each pairwise analysis is returned.
This data frame can be very large when there are many SNPs and many pairwise analyses, and the run-time of
\code{getIBDsegments()} will increase. The default is \code{posterior=FALSE}; \code{posterior=TRUE} is not recommended for large datasets.}
}
\value{
A named list of 1 object when \code{posterior=FALSE} and 2 objects when \code{posterior=TRUE}.
The first object in the list, \code{ibd_segments}, is a data frame with information:
\enumerate{
\item Family 1 ID (type \code{"character"})
\item Individual 1 ID (type \code{"character"})
\item Family 2 ID (type \code{"character"})
\item Individual 2 ID (type \code{"character"})
\item Chromosome (type \code{"numeric"} or \code{integer})
\item SNP identifier (type \code{"character"})
\item Start SNP (type \code{"character"})
\item End SNP (type \code{"character"})
\item Start position bp (type \code{"numeric"} or \code{integer})
\item End position bp (type \code{"numeric"} or \code{integer})
\item Start position M (type \code{"numeric"})
\item End position M (type \code{"numeric"})
\item Number of SNPs (type \code{"numeric"} or \code{integer})
\item Length bp (type \code{"numeric"} or \code{integer})
\item Length M (type \code{"numeric"})
\item IBD status (1 = one allele shared IBD, 2 = two alleles shared IBD) (type \code{"numeric"} or \code{integer})
}
where each row is a unique IBD segment for a pair of individuals. The data frame is headed
\code{fid1, iid1, fid2, iid2, chr, start.snp, end.snp, start.position.bp, end.position.bp, start.position.M, end.position.M,
number.snps, length.bp, length.M, ibd.status}. The second object (returned when \code{posterior=TRUE}), \code{posterior_probabilities}, is a
data frame with the first four columns
\enumerate{
\item Chromosome (type \code{"numeric"} or \code{integer})
\item SNP identifier (type \code{"character"})
\item Genetic map distance (type \code{"numeric"})
\item Base-pair positions (type \code{"numeric"} or \code{integer})
}
and columns 5 onwards are the posterior probabilities for each pair with pair identifier headers.
Rows correspond to SNPs.
}
\description{
Detects genomic regions shared IBD between pairs.
}
\examples{
\dontrun{
# infer IBD
my_ibd <- getIBDsegments(ped.genotypes = example_genotypes,
                         parameters = example_parameters,
                         model = NULL,
                         chromosomes = NULL,
                         number.cores = 1,
                         minimum.snps = 20,
                         minimum.length.bp = 50000,
                         error = 0.001,
                         posterior = FALSE)

str(my_ibd)
}
}
