% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identity_coefficients.R
\name{getIdentityCoef}
\alias{getIdentityCoef}
\title{Identity Coefficients}
\usage{
getIdentityCoef(pedigree, number.cores)
}
\arguments{
\item{pedigree}{a data frame containing a pedigree with 6 columns
\enumerate{
\item Family ID (type \code{"character"}, \code{"numeric"} or \code{"integer"})
\item Individual ID (\strong{\code{"numeric"}} or \strong{\code{"integer"}})
\item Paternal ID (\strong{\code{"numeric"}} or \strong{\code{"integer"}})
\item Maternal ID (\strong{\code{"numeric"}} or \strong{\code{"integer"}})
\item Gender (1 = male, 2 = female)
\item Phenotype (1 = unaffected, 2 = affected, 0 = unknown)
}
The family ID is not actually used in the analysis but is required for completeness of the pedigree;
thus \strong{all individual IDs must be unique and numeric.} \strong{The individuals in the pedigree
should also be numbered in a way such that every parent precedes his or her children.}}

\item{number.cores}{the number of cores used for parallel execution.}
}
\value{
A data frame with columns:
\enumerate{
\item Family 1 ID (type \code{"character"})
\item Individual 1 ID (type \code{"character"})
\item Family 2 ID (type \code{"character"})
\item Individual 2 ID (type \code{"character"})
\item The number of meiosis (type \code{"numeric"} or \code{"integer"})
\item Autosome identity coefficient for sharing 0 alleles IBD (type \code{"numeric"})
\item Autosome identity coefficient for sharing 1 allele IBD (type \code{"numeric"})
\item Autosome identity coefficient for sharing 2 alleles IBD (type \code{"numeric"})
\item X chromsome identity coefficient for sharing 0 alleles IBD (type \code{"numeric"})
\item X chromsome identity coefficient for sharing 1 allele IBD (type \code{"numeric"})
\item X chromsome identity coefficient for sharing 2 alleles IBD (type \code{"numeric"})
}
headed \code{fid1, iid1, fid2, iid2, a_i0, a_i1, a_i2, x_i0, x_i1} and \code{x_i2}, respectively.
Each row describes identity coefficient for a unique pair of samples.
}
\description{
Determines the identity coefficients based on a pedigree.
Autosome coefficients are calculated using the R package \code{identity} while
the X chromosome coefficients are implemented by \code{XIBD}.
}
\examples{
\dontrun{
# generate a pedigree
my_pedigree <- data.frame(fid = rep(1,16),
                          iid = 1:16,
                          pid = c(0,0,0,1,1,0,0,4,6,0,7,0,0,9,11,13),
                          mid = c(0,0,0,2,2,0,0,3,5,0,8,0,0,10,12,14),
                          sex = c(1,2,2,1,2,1,1,2,1,2,1,2,1,2,2,1),
                          aff = rep(1,16))

# calculate identity coefficients
identity_coef <- getIdentityCoef(pedigree = my_pedigree,
                                 number.cores = 1)

head(identity_coef)
}
}
