% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locus_proportions.R
\name{getLocusProportion}
\alias{getLocusProportion}
\title{XIBD Proportion of Pairs IBD}
\usage{
getLocusProportion(ped.genotypes, locus.matrix, groups = NULL)
}
\arguments{
\item{ped.genotypes}{a named list containing \code{pedigree}, \code{genotypes} and \code{model}.
See \code{Value} description in \code{\link{getGenotypes}} for more details.
The family IDs and individual IDs in \code{pedigree} must match the family IDs and individual IDs in the header of \code{genotypes}.}

\item{locus.matrix}{a data frame containing the binary IBD information for each SNP and each pair.
See \code{value} description in \code{\link{getLocusMatrix}} for more details.}

\item{groups}{a data frame with three columns of information:
\enumerate{
\item Family ID
\item Sample ID
\item Group ID
}
Group ID, for example, can be affection status.
If \code{groups} is specified, each sample in the pedigree should belong to a group,
and locus proportions will be calculated for each group.
The default is \code{groups=NULL}.}
}
\value{
A data frame the following 5 columns:
\enumerate{
\item Chromosome (type \code{"character"}, \code{"numeric"} or \code{"integer"})
\item SNP identifiers (type \code{"character"})
\item Genetic map distance (centi morgans, cM) (type \code{"numeric"})
\item Base-pair position (type \code{"integer"})
\item Proportion of pairs IBD (type \code{"integer"})
}
where each row describes a unique SNP. The data frame is headed
\code{chr, snp_id, pos_M, pos_bp} and \code{prop} respectively. If \code{groups} is
specified then there will be one column of proportions of each combination of groups with
these column header by group IDs and the number of pairs in the group.
}
\description{
Calculate the proportion of pairs IBD at each SNP.
}
\examples{
# generate a binary IBD matrix
my_locus_matrix <- getLocusMatrix(ped.genotypes = example_genotypes,
                                  ibd.segments = example_ibd)

# calculate the proportion of pairs IBD at each SNP
my_locus_prop <- getLocusProportion(ped.genotypes = example_genotypes,
                                    locus.matrix = my_locus_matrix,
                                    groups = NULL)
}
