% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hapmap_topbot.R
\docType{data}
\name{hapmap_topbot}
\alias{hapmap_topbot}
\title{Reference Topbot Data}
\format{Data frame containing
\describe{
\item{chr}{Chromosome (\code{"integer"}).}
\item{snp_id}{SNP identifier (type \code{"character"}).}
\item{pos_bp}{Base-pair position (type \code{"numeric"} or \code{"integer"}).}
\item{pos_M}{Genetic map distance (centi morgans, cM, or morgans, M) (type \code{"numeric"}).}
\item{TOPBOT}{Illuminas TOP or BOT designation of the SNP (type \code{"character"}).}
}}
\usage{
hapmap_topbot
}
\description{
The topbot annotation file containing Illumina TOP/BOT designation for the HapMap SNPs.
}
\keyword{datasets}

