% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_locus_proportions.R
\name{plotIBDproportions}
\alias{plotIBDproportions}
\title{XIBD Locus Proportion Plot}
\usage{
plotIBDproportions(locus.proportions, interval = NULL,
  annotation.genes = NULL, highlight.genes = NULL, add.rug = TRUE,
  plot.title = NULL)
}
\arguments{
\item{locus.proportions}{a data frame containing the proportion of pairs IBD at each SNP.
See \code{value} description in \code{\link{getLocusProportion}} for more details.}

\item{interval}{a vector of length 3 containing a chromosome, a start position (bp) and an end position (bp)
of an interval to plot. The default is \code{interval=NULL} which will plot the proportions over all chromosomes
in \code{locus.proportions}.}

\item{annotation.genes}{a data frame with at least 5 columns of information:
\enumerate{
\item Chromosome (type \code{"numeric"} or \code{"integer"})
\item Gene name (type \code{"character"})
\item Start location of the gene in base-pairs (type \code{"numeric"} or \code{"integer"})
\item End location of the gene in base-pairs (type \code{"numeric"} or \code{"integer"})
\item Gene strand (+ or -)
}
\code{annotation.genes} should contain the following headers \code{chr, name, start, end} and \code{strand}.
This data frame does not have to be in a specific order, however it must contain all of the above information
with respective labels.}

\item{highlight.genes}{a data frame with at least 4 columns of information:
\enumerate{
\item Chromosome (type \code{"numeric"} or \code{"integer"})
\item Gene name (type \code{"character"})
\item Start location of the gene in base-pairs (type \code{"numeric"} or \code{"integer"})
\item End location of the gene in base-pairs (type \code{"numeric"} or \code{"integer"})
}
\code{highlight.genes} should contain the following headers \code{chr, name, start} and \code{end}.
This data frame does not have to be in a specific order, however it must contain all of the above information
with respective labels.}

\item{plot.title}{a character string of a title to be added to the plot. The default is \code{plot.title=NULL} which does not add a title to the plot.}

\item{add.legend}{a logical value indicating whether the IBD status legend should be plotted. The default is \code{add.legend=TRUE}.}
}
\description{
Plot the proportion of pairs IBD for each SNP across the genome.
Annotation genes can be added to the plot and specific genes on interest can be highlighted.
}

