% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ibd_segments.R
\name{plotIBDsegments}
\alias{plotIBDsegments}
\title{XIBD Plot IBD Segments}
\usage{
plotIBDsegments(ibd.segments, ped.genotypes, interval = NULL,
  annotation.genes = NULL, highlight.genes = NULL, segment.height = 0.5,
  number.per.page = NULL, add.fid.name = TRUE, add.iid.name = TRUE,
  add.rug = FALSE, plot.title = NULL, add.legend = TRUE)
}
\arguments{
\item{ibd.segments}{a named list containing the \code{ibd_segments} inferred from pairs of samples.
See \code{value} description in \code{\link{getIBDsegments}} for more details.}

\item{ped.genotypes}{a named list containing \code{pedigree}, \code{genotypes} and \code{model}.
See \code{Value} description in \code{\link{getGenotypes}} for more details.
The family IDs and individual IDs in \code{pedigree} must match the family IDs and individual IDs in the header of \code{genotypes}.}

\item{interval}{a vector of length 3 containing a chromosome, a start position (bp) and an end position (bp)
of an interval to plot. The default is \code{interval=NULL} which will plot the IBD segments over all chromosomes
in \code{ped.genotypes}.}

\item{annotation.genes}{a data frame with at least 5 columns of information:
\enumerate{
\item Chromosome (type \code{numeric} or \code{integer})
\item Gene name (type \code{character})
\item Start location of the gene in base-pairs (type \code{numeric} or \code{integer})
\item End location of the gene in base-pairs (type \code{numeric} or \code{integer})
\item Gene strand (+ or -)
}
\code{annotation.genes} should contain the following headers \code{chr, name, start, end} and \code{strand}.
This data frame does not have to be in a specific order, however it must contain all of the above information
with respective labels.}

\item{highlight.genes}{a data frame with at least 4 columns of information:
\enumerate{
\item Chromosome (type \code{numeric} or \code{integer})
\item Gene name (type \code{character})
\item Start location of the gene in base-pairs (type \code{numeric} or \code{integer})
\item End location of the gene in base-pairs (type \code{numeric} or \code{integer})
}
\code{highlight.genes} should contain the following headers \code{chr, name, start} and \code{end}.
This data frame does not have to be in a specific order, however it must contain all of the above information
with respective labels.}

\item{segment.height}{the height of IBD segment blocks, such that 0 < \code{segment.height} <= 1. The default is \code{segment.height=0.5}.}

\item{number.per.page}{the maximum number of IBD pairs to plot in a single praphics window. The default is
\code{number.per.page=NULL} which will plot all IBD pairs in a single window. This may not be ideal when there are many IBD pairs. If
\code{number.per.page} is set, it is recommended to plot the output to a file as opposed to the R console.}

\item{add.fid.name}{a logical value indicating if family IDs should be included in the y-axis labels.
The default is \code{add.fid.name=TRUE}.}

\item{add.iid.name}{a logical value indicating if individual IDs should be included in the y-axis labels.
The default is \code{add.iid.name=TRUE}.}

\item{add.rug}{a logical value indicating whether SNP positions should be added to the plot. The default is \code{add.rug=TRUE}.}

\item{plot.title}{a character string of a title to be added to the plot. The default is \code{plot.title=NULL} which does not add a title to the plot.}

\item{add.legend}{a logical value indicating whether the IBD status legend should be plotted. The default is \code{add.legend=TRUE}.}
}
\description{
Plot IBD segments for pairs across the genome.
Annotation genes can be added to the plot and specific genes on interest can be highlighted.
}
\examples{
# plot IBD segments
plotIBDsegments(ibd.segments = example_ibd,
                ped.genotypes = example_genotypes,
                interval = NULL,
                annotation.genes = NULL,
                highlight.genes = NULL,
                segment.height = 0.5,
                number.per.page = NULL,
                add.fid.name = FALSE,
                add.iid.name = TRUE,
                add.rug = TRUE,
                plot.title = "Inferred IBD Segments",
                add.legend = TRUE)
}
