% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_pedmap.R
\docType{data}
\name{reference_pedmap}
\alias{reference_pedmap}
\title{Subset of Reference Data}
\format{List with two objects
\describe{
\item{PED}{PLINK PED data frame. See \code{\link{getGenotypes}} for details.}
\item{MAP}{PLINK MAP data frame. See \code{\link{getGenotypes}} for details.}
}}
\usage{
reference_pedmap
}
\description{
A subset of the HapMap TSI haplotype data in PLINK PED/MAP format
for chromosomes 1-23.
The data is for 102 samples and 28,808 SNPs.
}
\keyword{datasets}

