% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_topbot.R
\docType{data}
\name{reference_topbot}
\alias{reference_topbot}
\title{Reference TOPBOT Data}
\format{See \code{\link{switchBOTgenotypes}} for details.}
\usage{
reference_topbot
}
\description{
A subset of the topbot annotation file found at
\url{http://bioinf.wehi.edu.au/software/XIBD}
}
\keyword{datasets}

