% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_bot_genotypes.R
\name{switchBOTgenotypes}
\alias{switchBOTgenotypes}
\title{XIBD BOT Genotype Switching}
\usage{
switchBOTgenotypes(ped.genotypes, topbot.annotation)
}
\arguments{
\item{ped.genotypes}{a named list containing \code{pedigree}, \code{genotypes} and \code{model}.
See \code{Value} description in \code{\link{getGenotypes}} for more details.
The family IDs and individual IDs in \code{pedigree} must match the family IDs and individual IDs in the header of \code{genotypes}.}

\item{topbot.annotation}{the annotation file provided by XIBD containing information of TOPBOT status for each SNP
and population allele frequencies of the 11 HapMap populations.
This file can be downloaded from \url{http://bioinf.wehi.edu.au/software/XIBD}.}
}
\value{
A named list of the same format as the input \code{ped.genotypes} with A and B alleles switched for BOT SNPs.
}
\description{
The HapMap allele frequencies in XIBDs HapMap allele frequency files are calculated for the A allele only,
where the A allele is determined by the following rules:
\enumerate{
\item When one of the possible variations of the SNP is adenine (A), then adenine is labeled the A allele
and the remaining variation is labeled the B allele, regardless of what this might be.
\item If adenine (A) is not a variation of the SNP but cytosine (C) is, then cytosine is labeled the A allele
and the remaining variation is labeled the B allele.
\item If neither adenine (A) or cytosine (C) are variants of the SNP then thymine (T) is labeled the A allele.
}
Illuminas convention for the naming of A and B alleles differs to that of the HapMap data
(http://www.illumina.com/documents/products/technotes/technote_topbot.pdf). Rather, the classification
of A and B alleles depend on the top (TOP) and bottom (BOT) designations of the SNP. This
means that the A allele in the HapMap data is not always the same as the A allele in the Illumina data. In
fact, alleles that have been named according to the BOT designation actually correspond the the B allele
in the HapMap data. To correct for this, \code{switchBOTgenotypes()} switchs the A and B alleles in
the input genotypes for all SNPs corresponding to BOT designations. This mean a homozygous genotype, 0, will be
changed to a homozygous alternative genotype, 2, and vis versa. Heterozygous genotypes will be unchanged.
NOTE: this function should only be implemented with Illumina SNPchip data when XIBD's HapMap reference data is used
and if there is a noticeable discrepancy between population allele frequencies calculated from the HapMap reference data
and those calculated from the input dataset.
}

